# tierkreis_visualization/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/workflows/": {"get": {"summary": "List Workflows", "operationId": "list_workflows_api_workflows__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/WorkflowDisplay"}, "type": "array", "title": "Response List Workflows Api Workflows  Get"}}}}}}}, "/api/workflows/{workflow_id}/graphs": {"get": {"summary": "List Nodes", "operationId": "list_nodes_api_workflows__workflow_id__graphs_get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "locs", "in": "query", "required": true, "schema": {"type": "array", "items": {"type": "string"}, "title": "Locs"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GraphsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/nodes/{node_location_str}": {"get": {"summary": "Get Node", "operationId": "get_node_api_workflows__workflow_id__nodes__node_location_str__get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "node_location_str", "in": "path", "required": true, "schema": {"type": "string", "title": "Node Location Str"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PyGraph"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/nodes/{node_location_str}/outputs": {"get": {"summary": "Get Eval Outputs", "operationId": "get_eval_outputs_api_workflows__workflow_id__nodes__node_location_str__outputs_get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "node_location_str", "in": "path", "required": true, "schema": {"type": "string", "title": "Node Location Str"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/nodes/{node_location_str}/inputs/{port_name}": {"get": {"summary": "Get Input", "operationId": "get_input_api_workflows__workflow_id__nodes__node_location_str__inputs__port_name__get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "node_location_str", "in": "path", "required": true, "schema": {"type": "string", "title": "Node Location Str"}}, {"name": "port_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Port Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/nodes/{node_location_str}/outputs/{port_name}": {"get": {"summary": "Get Output", "operationId": "get_output_api_workflows__workflow_id__nodes__node_location_str__outputs__port_name__get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "node_location_str", "in": "path", "required": true, "schema": {"type": "string", "title": "Node Location Str"}}, {"name": "port_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Port Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/logs": {"get": {"summary": "Get Logs", "operationId": "get_logs_api_workflows__workflow_id__logs_get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/workflows/{workflow_id}/nodes/{node_location_str}/errors": {"get": {"summary": "Get Errors", "operationId": "get_errors_api_workflows__workflow_id__nodes__node_location_str__errors_get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}, {"name": "node_location_str", "in": "path", "required": true, "schema": {"type": "string", "title": "Node Location Str"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/{path}": {"get": {"summary": "Read Root", "operationId": "read_root__path__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"GraphsResponse": {"properties": {"graphs": {"additionalProperties": {"$ref": "#/components/schemas/PyGraph"}, "type": "object", "title": "Graphs"}}, "type": "object", "required": ["graphs"], "title": "GraphsResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "PyEdge": {"properties": {"from_node": {"type": "string", "title": "From Node"}, "from_port": {"type": "string", "title": "From Port"}, "to_node": {"type": "string", "title": "To Node"}, "to_port": {"type": "string", "title": "To Port"}, "value": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Value"}, "conditional": {"type": "boolean", "title": "Conditional", "default": false}}, "type": "object", "required": ["from_node", "from_port", "to_node", "to_port"], "title": "PyEdge"}, "PyGraph": {"properties": {"nodes": {"items": {"$ref": "#/components/schemas/PyNode"}, "type": "array", "title": "Nodes"}, "edges": {"items": {"$ref": "#/components/schemas/PyEdge"}, "type": "array", "title": "Edges"}}, "type": "object", "required": ["nodes", "edges"], "title": "PyGraph"}, "PyNode": {"properties": {"id": {"type": "string", "title": "Id"}, "status": {"type": "string", "enum": ["Not started", "Started", "Error", "Finished"], "title": "Status"}, "function_name": {"type": "string", "title": "Function Name"}, "node_type": {"type": "string", "enum": ["function", "ifelse", "map", "eval", "loop", "eifelse", "const", "output", "input"], "title": "Node Type"}, "node_location": {"type": "string", "title": "Node Location", "default": ""}, "outputs": {"items": {"type": "string"}, "type": "array", "title": "Outputs"}, "value": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Value"}, "started_time": {"type": "string", "title": "Started Time"}, "finished_time": {"type": "string", "title": "Finished Time"}}, "type": "object", "required": ["id", "status", "function_name", "node_type", "outputs", "started_time", "finished_time"], "title": "PyNode"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WorkflowDisplay": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "id_int": {"type": "integer", "title": "Id Int"}, "name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "start_time": {"type": "string", "title": "Start Time"}, "errors": {"items": {"type": "string"}, "type": "array", "title": "Errors"}}, "type": "object", "required": ["id", "id_int", "name", "start_time", "errors"], "title": "WorkflowDisplay"}}}}